Upgrade ChicagoBoss to version 0.8.13 to support Erlang/OTP 17
This also changes the minimum required OTP version for the web ui to
R16B03-1. The README.md file is updated for this. ChicagoBoss does not
work with R16B03 because of a bug in that version. ChicagoBoss is
documented to work with R15B03* but ChicagoBoss 0.8.13 does not
compile with R15B03.

https://github.com/ChicagoBoss/ChicagoBoss/wiki/Quickstart/f8f1d2d67867953190608b0f6d39e6853df75d86