Add R-D-IM app and trivial benchmark that works
Some files from the original app were removed because bencherl didn't need
them, e.g. README.md, start_gui.sh, start.sh. The nodes of the complete
architecture are deployed by bencherl (in the same host as in the vanilla
R-D-IM app) as described in conf/run.conf.

A few more changes include passing a directory path where the .csv output files
should be stored (this was implemented by passing one more argument in
logger:launch_latency) and sending a message to a process coordinator (spawned
in the master or dashboard node) to notify that a logger has stopped. In that
way, the im_bench benchmark can be sure that when all loggers stopped the
benchmarking is over (this is because there is no logger:stop function provided
by the vanilla app).